+++ tycho-its/src/test/java/org/eclipse/tycho/test/pomDependencyConsider/LocalMavenRepositoryTool.java
@@ -26,8 +26,8 @@ import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-import org.eclipse.tycho.locking.facade.FileLockService;
-import org.eclipse.tycho.locking.facade.FileLocker;
+import org.eclipse.tycho.FileLockService;
+import org.eclipse.tycho.FileLocker;
 import org.eclipse.tycho.test.util.EnvironmentUtil;
 
 public class LocalMavenRepositoryTool {